get display-p3 borders working in Android

+++ packages/react-native/ReactAndroid/src/main/java/com/facebook/react/uimanager/BaseViewManager.java
@@ -167,7 +167,7 @@ public abstract class BaseViewManager<T extends View, C extends LayoutShadowNode
 
   @Override
   public void setBackgroundColor(@NonNull T view, int backgroundColor) {
-    view.setBackgroundColor((int)backgroundColor);
+    view.setBackgroundColor(backgroundColor);
   }
 
   @ReactProp(
